Fix Java distance formula in function 01
The current implementations of the distance method between the Java and
JavaScript versions of function 01_Intro_ProcessLargeData are currently
calculating vastly different distances with the same inputs.

The JavaScript version of the function is more accurate based on a
Google Maps comparison.

* Modify the distance method in the Java function to calculate the
  distance the same way as the JavaScript distance method.
* Add assertions about the accuracy of the distance to the tests for
  both the JavaScript and Java tests that have non-empty payloads.

Fixes #707 01_Intro_ProcessLargeData_Java does not calculate distance
correctly